✨ feat: Add filter logic based on bookmarked scholarships
This commit adds logic to filter scholarships based on whether they are bookmarked by the current user. If the filter parameter is set to 'bookmarked' and there is a current user, only scholarships that have been bookmarked by the user will be displayed. Additionally, a visual indicator is added to show the active state of the bookmarked filter button in the UI.